# Resolve variables and functions statically