--- a/src/prelude.rs
+++ b/src/prelude.rs
@@ -63,8 +63,8 @@ impl Prelude {
     pub fn from_module(module: &Module) -> Prelude {
         let mut functions = HashMap::new();
         intrinsics::standard(&mut functions);
-        for (key, &(_, ref val)) in &*module.ext_prelude {
-            functions.insert(key.clone(), val.clone());
+        for f in &*module.ext_prelude {
+            functions.insert(f.name.clone(), f.p.clone());
         }
         for f in &module.functions {
             functions.insert(f.name.clone(), PreludeFunction::new(f));