Closes https://github.com/PistonDevelopers/dyon/issues/51

### What is statically resolving?

When you have a function like this:

```rust
fn foo() {
    x := 5 // <--- declaring `x`
   println(x) // <-- referencing `x`
}
```

Previously, the resolving of references happened by pushing local variables to its own stack and then look up the name in a loop. Functions were looked up through a `HashMap`.

With statically resolving, the references stores an index in `ast::Item` that points directly to the stack location relative to the end of the stack using the formula `end - static_id`.

This requires that the stack memory layout is consistent each time the program reaches the same reference.

The benefit of statically resolving is that there is less overhead, and easier to do other things like generating Rust code.

### Default Cargo feature: debug_resolve

Added a default Cargo feature "debug_resolve". This adds some overhead, but helps with debugging the static resolver. To benchmark, use `cargo bench --no-default-features`.

There are two things that can go wrong with statically resolving:

1. The statically resolver does not use same stack memory layout as the run-time
2. The run-time memory layout is inconsistent when executing

1) is checked for by comparing a run-time resolved value against the statically resolved one.

2) is checked for by looking up the variable in the `local_stack` and compared against the resolved one.

If 1) or 2) are wrong, a run-time error will be created.

"debug_resolve" will be enabled by default until the language stabilizes.

### Performance

The performance is slightly better overall.

```
Before:
test tests::bench_add           ... bench:  43,689,453 ns/iter (+/- 5,008,275)
test tests::bench_add_n         ... bench:  16,288,823 ns/iter (+/- 3,310,947)
test tests::bench_array         ... bench:  19,802,667 ns/iter (+/- 3,726,364)
test tests::bench_call          ... bench:  18,976,300 ns/iter (+/- 4,034,740)
test tests::bench_len           ... bench:   4,156,717 ns/iter (+/- 886,832)
test tests::bench_main          ... bench:   2,483,362 ns/iter (+/- 295,884)
test tests::bench_min           ... bench:  18,103,155 ns/iter (+/- 2,930,421)
test tests::bench_min_fn        ... bench:  88,207,328 ns/iter (+/- 9,629,365)
test tests::bench_n_body        ... bench:  80,031,759 ns/iter (+/- 7,977,068)
test tests::bench_object        ... bench:  27,863,104 ns/iter (+/- 7,286,857)
test tests::bench_primes        ... bench:  50,798,333 ns/iter (+/- 9,688,478)
test tests::bench_primes_trad   ... bench:  48,398,347 ns/iter (+/- 7,762,432)
test tests::bench_sum           ... bench:   7,729,297 ns/iter (+/- 1,419,093)
test tests::bench_threads_go    ... bench:  38,976,704 ns/iter (+/- 4,364,484)
test tests::bench_threads_no_go ... bench:  47,215,235 ns/iter (+/- 7,519,226)

After:
test tests::bench_add           ... bench:  41,542,169 ns/iter (+/- 4,234,672)
test tests::bench_add_n         ... bench:  16,261,309 ns/iter (+/- 3,834,949)
test tests::bench_array         ... bench:  19,255,340 ns/iter (+/- 3,055,505)
test tests::bench_call          ... bench:  14,328,926 ns/iter (+/- 2,317,606)
test tests::bench_len           ... bench:   4,104,587 ns/iter (+/- 618,239)
test tests::bench_main          ... bench:   2,501,248 ns/iter (+/- 190,775)
test tests::bench_min           ... bench:  17,424,805 ns/iter (+/- 3,484,824)
test tests::bench_min_fn        ... bench:  73,552,096 ns/iter (+/- 6,110,614)
test tests::bench_n_body        ... bench:  79,849,883 ns/iter (+/- 5,914,737)
test tests::bench_object        ... bench:  28,193,145 ns/iter (+/- 5,410,740)
test tests::bench_primes        ... bench:  49,807,590 ns/iter (+/- 6,107,639)
test tests::bench_primes_trad   ... bench:  46,189,252 ns/iter (+/- 5,943,564)
test tests::bench_sum           ... bench:   7,570,734 ns/iter (+/- 763,915)
test tests::bench_threads_go    ... bench:  35,650,443 ns/iter (+/- 3,320,686)
test tests::bench_threads_no_go ... bench:  44,099,258 ns/iter (+/- 8,645,742)
```

### Future work

The local stack is still used, but can be left out when "debug_resolve" is not enabled. This might improve performance a bit.